chatbot-service/requirements.txt
mysql-connector-python
pymongo
openai
chromadb
pandas
numpy